+++ autd3-driver/src/defined/mod.rs
@@ -2,15 +2,15 @@ mod angle;
 mod control_point;
 mod freq;
 
-pub use std::f64::consts::PI;
+pub use std::f32::consts::PI;
 
 #[cfg(feature = "use_meter")]
 mod unit {
-    pub const METER: f64 = 1.0;
+    pub const METER: f32 = 1.0;
 }
 #[cfg(not(feature = "use_meter"))]
 mod unit {
-    pub const METER: f64 = 1000.0;
+    pub const METER: f32 = 1000.0;
 }
 pub use unit::*;
 
@@ -18,17 +18,17 @@ pub use angle::*;
 pub use control_point::*;
 pub use freq::*;
 
-pub const MILLIMETER: f64 = METER / 1000.0;
+pub const MILLIMETER: f32 = METER / 1000.0;
 
-pub type Complex = nalgebra::Complex<f64>;
+pub type Complex = nalgebra::Complex<f32>;
 
-pub const ABSOLUTE_THRESHOLD_OF_HEARING: f64 = 20e-6;
+pub const ABSOLUTE_THRESHOLD_OF_HEARING: f32 = 20e-6;
 
-pub const T4010A1_AMPLITUDE: f64 = 275.574246625 * 200.0 * MILLIMETER; // [Pa*mm]
+pub const T4010A1_AMPLITUDE: f32 = 275.574246625 * 200.0 * MILLIMETER; // [Pa*mm]
 
 pub const DEFAULT_TIMEOUT: std::time::Duration = std::time::Duration::from_millis(200);
 
 pub const FREQ_40K: Freq<u32> = Freq { freq: 40000 };
 
 #[allow(non_upper_case_globals)]
-pub const mm: f64 = MILLIMETER;
+pub const mm: f32 = MILLIMETER;
